requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
prophet==1.1.4
statsmodels==0.14.0
keras==2.15.0

# Technical Analysis
TA-Lib==0.4.28
ta==0.10.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Report generation
jinja2==3.1.2
weasyprint==60.2
reportlab==4.0.7

# Web scraping and sentiment
beautifulsoup4==4.12.2
nltk==3.8.1
textblob==0.17.1
newspaper3k==0.2.8

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
